a2a-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a2a-server"
version = "0.1.0"
description = "Example A2A server with OpenAI-powered chat agent"
authors = [
    { name = "assistant-ui Team", email = "team@assistant-ui.com" }
]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "openai>=1.0.0",
    "uvicorn>=0.34.0",
    "starlette>=0.46.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
